test: consolidate all event bus tests into single file
Merged event_bus_extended_test.go and event_bus_context_any_test.go
into event_bus_test.go for better organization while maintaining
100% test coverage.
